third_party/blink/web_tests/external/wpt/digital-credentials/create.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] navigator.credentials.create() promise is rejected if called with an aborted controller.
  promise_rejects_dom: function "function() { throw e; }" threw object "NotSupportedError: Only exactly one of 'password', 'federated', and 'publicKey' credential types are currently supported." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] navigator.credentials.create() promise is rejected if called with an aborted controller in same-origin iframe.
  promise_rejects_dom: function "function() { throw e; }" threw object "NotSupportedError: Only exactly one of 'password', 'federated', and 'publicKey' credential types are currently supported." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] navigator.credentials.create() promise is rejected if called with an aborted signal in cross-origin iframe.
  assert_equals: expected "AbortError" but got "NotSupportedError"
[FAIL] navigator.credentials.create() promise is rejected if abort controller is aborted after call to create().
  promise_rejects_dom: function "function() { throw e; }" threw object "TypeError: Failed to execute 'create' on 'CredentialsContainer': Failed to read the 'digital' property from 'CredentialCreationOptions': Failed to read the 'data' property from 'DigitalCredentialCreationOptions': Required member is undefined." that is not a DOMException AbortError: property "code" is equal to undefined, expected 20
[FAIL] navigator.credentials.create() promise is rejected if abort controller is aborted after call to create() in cross-origin iframe.
  assert_equals: expected "DOMException" but got "TypeError"
[FAIL] Adding mediations together with abort signal respects the abort signal.
  promise_rejects_dom: function "function() { throw e; }" threw object "NotSupportedError: Only exactly one of 'password', 'federated', and 'publicKey' credential types are currently supported." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
Harness: the test ran to completion.
